Mark Level as serde Serialize-able
RFC!

The reason to do this is to use this struct in the `Line` struct under
`servo/components/layout/`, which is `Serialize`-able already and uses
`u8` at the moment.

So, do you think this big dependency worth it by itself?

If not, how about have conditional `#derive` with module availablity, so
whenever serve+serde_derive are available, also make it
`Serialize`-able, otherwise not.

Or, should use serde's "remote derive" in `servo/components/layout/` and
not bother about it here?