Create Bidi Level struct type
The `Level` struct always represents a valid bidi embedding level, with
checks for explicit and implicit boundries (max_depth and max_depth+1,
respectively) therefore can fail on `new*()` and mutation
(`raise*()`/`lower()`) methods.

So, we have one level type, which is always a valid number of *implicit
embedding level resolution*, but can be made invalid for *explicit
embedding level resolution*. This is fine because external usage of the
data structure would normally deal with the final resolved levels, which
are the *implicit* ones (and that's why its boundry is made the implicit
boundry here, as well), and the *explicit* bounrdy is supposed to be
checked only during the early stages of the algorithm (the explicit
stages), which can do so explicitly by calling the appropriate methods.