Add serde support and bump version to 0.3.0
Add `with_serde` feature, which implements serde for the new `struct
Level`, mainly used in `servo`, supporting serde `0.8`, `0.9` and `1.0`.

Add tests for the `with_serde` feature.  The `serde_tests` modules only
works for `serde:>=1.0`, though.

`servo` has dependency on the loose implementation of `visual_runs()`,
which couldn not be improved without breaking the API, as it needs more
information to process the levels correctly, and the call-site in
`servo` does not have all the information needed and needs a non-trivial
change to work with the new improved version. Therefore, I have moved the
old version to a `deprecated` module, to be used for now until `servo`
is fixed and we drop the old implementation.

Bump version to `0.3.0`, as we now ready for a release: can build
`servo` (patch ready) and `idna` crates.