[benches/udhr] Run reorder_line() on all paragraphs
Previously, the reorder logic was only called on the first paragraph,
which is an ASCII-only line on all files, having no bidi content,
therefore not measuring the optimizations for LTR-only text. Now fixing
the benchmarks to run on all text, to fix this problem.

Also, isolate benchmark for `reorder_line()`.

And, put 3 misplaces files in the correct group.